+++ lib/src/xml_events/codec/node_codec.dart
@@ -1,6 +1,6 @@
 import 'dart:convert' show Codec, Converter;
 
-import '../../../xml.dart' show XmlNode;
+import '../../xml/nodes/node.dart';
 import '../converters/node_decoder.dart';
 import '../converters/node_encoder.dart';
 import '../event.dart';