+++ lib/src/xml_events/events/doctype.dart
@@ -1,10 +1,10 @@
-import '../../../xml.dart' show XmlNodeType;
+import '../../xml/utils/node_type.dart';
 import '../event.dart';
 import '../visitor.dart';
 
 /// Event of an XML doctype node.
 class XmlDoctypeEvent extends XmlEvent {
-  const XmlDoctypeEvent(this.text);
+  XmlDoctypeEvent(this.text);
 
   final String text;